Fix minor TOC formatting issue.
Some markdown processors don't like having an HTML comment immediately
before or after any markdown element.

This change adds a blank line before and after the

<!-- START TOC -->

and

<!-- END TOC -->

HTML comment elements, and normalizes the file so that the number of
blank lines around the TOC remains constant when the TOC is stripped out
and re-added.